Extract engine helpers into modules with tests
